move _ImageHander and expose publicly, add glitched crate
